requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
tensorflow==2.8.0
numpy==1.21.2
Pillow==8.3.2
python-multipart==0.0.5 